Add comments, Refactore
