+++ src/app/providers/StoreProvider/config/hooks.ts
@@ -1,4 +1,5 @@
 import { useDispatch } from "react-redux";
 import { type AppDispatch } from "./store";
 
+// Создаем кастомный диспатч хук, чтобы пользоваться redux-thunk
 export const useAppDispatch: () => AppDispatch = useDispatch;
