+++ src/app/providers/StoreProvider/config/store.ts
@@ -20,7 +20,7 @@ export default function configureStore() {
     const store = createStore(
         rootReducer,
         {},
-        composeWithDevTools(middlewareEnhancer)
+        composeWithDevTools(middlewareEnhancer) // дев тулсы, чтобы было удобно отслеживать изменения
     );
 
     return store;
