+++ src/features/getReviewer/ui/getReviewer/GetReviewer.tsx
@@ -1,15 +1,11 @@
-import { getRandomReviewer } from "../../model/services/getRandomReviewer";
-import { GithubUserDataType } from "../../model/types/types";
-import { useContext, useState } from "react";
-import { SETTINGS_LOCALSTORAGE_KEY } from "shared/const/localStorage";
+import { useContext } from "react";
 import { classNames } from "shared/lib/classNames/classNames";
-import { setLocalStorageItemSafe } from "shared/lib/localStorage/localStorageSafe";
 import { SettingForm } from "../settingForm/SettingForm";
 import { UsersInfo } from "../usersInfo/UsersInfo";
 import cls from "./GetReviewer.module.css";
 import { LocalStorageContext } from "app/providers/LocalStorageProvider";
-import { useDispatch, useSelector } from "react-redux";
-import { getRandomReviewerData } from "../../model/services/getRandomReviewerData";
+import { useSelector } from "react-redux";
+import { getRandomReviewerData } from "../../model/actionCreators/getRandomReviewerData";
 import { useAppDispatch } from "app/providers/StoreProvider";
 import {
     getGetReviewerError,
@@ -30,52 +26,16 @@ export const GetReviewer = ({ className }: GetReviewerProps) => {
     const owner = useSelector(getGetReviewerOwner);
     const reviewer = useSelector(getGetReviewerReviewer);
     const restContributors = useSelector(getGetReviewerRest);
-    // const [error, setError] = useState<string>("");
-    // const [loading, setLoading] = useState<boolean>(false);
-    // const [owner, setOwner] = useState<null | GithubUserDataType>(null);
-    // const [reviewer, setReviewer] = useState<null | GithubUserDataType>(null);
-    // const [restContributors, setRestContributors] = useState<
-    //     Array<GithubUserDataType>
-    // >([]);
 
     const isLocalStorage = useContext(LocalStorageContext);
 
-    // callback который будет срабатывать на отправку формы настроек. Получает рандомного ревьювера, сохраняет данные
+    // callback который будет срабатывать на отправку формы настроек. Диспатчит асинхронный экшн, который в свою очередь дергает ручку, сохраняет данные в стор и в LS
     const onSubmitForm = async (
         login: string,
         repo: string,
         blacklist: string
     ) => {
         dispatch(getRandomReviewerData(login, repo, blacklist, isLocalStorage));
-        // setLoading(true);
-        // setError("");
-
-        // try {
-        //     // Получаем рандомного ревьювера
-        //     const { owner, reviewer, rest } = await getRandomReviewer(
-        //         login,
-        //         repo,
-        //         blacklist
-        //     );
-        //     // Сохраняем данные в стейт
-        //     owner
-        //         ? setOwner(owner)
-        //         : setOwner({ login, contributions: 0, type: "User" });
-        //     setReviewer(reviewer);
-        //     setRestContributors(rest);
-
-        //     // Сохраняем данные в LocalStorage
-        //     if (isLocalStorage) {
-        //         setLocalStorageItemSafe(SETTINGS_LOCALSTORAGE_KEY, {
-        //             login,
-        //             repo,
-        //             blacklist,
-        //         });
-        //     }
-        // } catch (error) {
-        //     setError(error as string);
-        // }
-        // setLoading(false);
     };
 
     return (
